Update token retrieval instructions in install docs
The way the token is retrieved changed starting with Kubernetes v1.22.
This was updated elsewhere, but the install documentation on the website
still only had the pre-v1.22 instructions. If someone were to just
follow the instructions from the site they wouldn't be able to retrieve
the token necessary to log in.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>